Modify GeoJSON to use `Location` instead of `CLLocation`
Extract `toCoreLocation()` method from `GeoJSONMessage` to separate extension
